Completed 4, 5 and 6.
5 and 6 were easy - 4 made me realise there was an issue with the
compiler in Eclipse which was 1) massively slowing down the program
speed and 2) not converting ints to strings, creating gibberish and
breaking the calc.

Updating the C++ compiler and ensuring we used the most up-to-date
conversion cut program time for 4 down from 10s of minutes down into the
milliseconds (microseconds?).

I'm enjoying learning C++, it's great that I can do some things with
relative ease. But so far most of my errors have been completely
mythological and unapparent - debugging is a nightmare.